Entrega Final

+++ src/controladores/VendedoresJpaController.java
@@ -5,17 +5,20 @@
  */
 package controladores;
 
+import controladores.exceptions.IllegalOrphanException;
 import controladores.exceptions.NonexistentEntityException;
 import entidades.Vendedores;
 import java.io.Serializable;
+import javax.persistence.Query;
+import javax.persistence.EntityNotFoundException;
+import javax.persistence.criteria.CriteriaQuery;
+import javax.persistence.criteria.Root;
+import entidades.Ventas;
+import java.util.ArrayList;
 import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
-import javax.persistence.Query;
-import javax.persistence.EntityNotFoundException;
 import javax.persistence.Persistence;
-import javax.persistence.criteria.CriteriaQuery;
-import javax.persistence.criteria.Root;
 
 /**
  *
@@ -26,9 +29,11 @@ public class VendedoresJpaController implements Serializable {
     public VendedoresJpaController(EntityManagerFactory emf) {
         this.emf = emf;
     }
-     public VendedoresJpaController() {
-        this.emf = Persistence.createEntityManagerFactory("ProyectoOriginalPU");
+
+    public VendedoresJpaController() {
+         this.emf = Persistence.createEntityManagerFactory("ProyectoOriginalPU");
     }
+    
     private EntityManagerFactory emf = null;
 
     public EntityManager getEntityManager() {
@@ -36,11 +41,29 @@ public class VendedoresJpaController implements Serializable {
     }
 
     public void create(Vendedores vendedores) {
+        if (vendedores.getVentasList() == null) {
+            vendedores.setVentasList(new ArrayList<Ventas>());
+        }
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            List<Ventas> attachedVentasList = new ArrayList<Ventas>();
+            for (Ventas ventasListVentasToAttach : vendedores.getVentasList()) {
+                ventasListVentasToAttach = em.getReference(ventasListVentasToAttach.getClass(), ventasListVentasToAttach.getId());
+                attachedVentasList.add(ventasListVentasToAttach);
+            }
+            vendedores.setVentasList(attachedVentasList);
             em.persist(vendedores);
+            for (Ventas ventasListVentas : vendedores.getVentasList()) {
+                Vendedores oldIdVendedorOfVentasListVentas = ventasListVentas.getIdVendedor();
+                ventasListVentas.setIdVendedor(vendedores);
+                ventasListVentas = em.merge(ventasListVentas);
+                if (oldIdVendedorOfVentasListVentas != null) {
+                    oldIdVendedorOfVentasListVentas.getVentasList().remove(ventasListVentas);
+                    oldIdVendedorOfVentasListVentas = em.merge(oldIdVendedorOfVentasListVentas);
+                }
+            }
             em.getTransaction().commit();
         } finally {
             if (em != null) {
@@ -49,12 +72,46 @@ public class VendedoresJpaController implements Serializable {
         }
     }
 
-    public void edit(Vendedores vendedores) throws NonexistentEntityException, Exception {
+    public void edit(Vendedores vendedores) throws IllegalOrphanException, NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
             vendedores = em.merge(vendedores);
+            Vendedores persistentVendedores = em.find(Vendedores.class, vendedores.getId());
+            List<Ventas> ventasListOld = persistentVendedores.getVentasList();
+            List<Ventas> ventasListNew = vendedores.getVentasList();
+            List<String> illegalOrphanMessages = null;
+            for (Ventas ventasListOldVentas : ventasListOld) {
+                if (!ventasListNew.contains(ventasListOldVentas)) {
+                    if (illegalOrphanMessages == null) {
+                        illegalOrphanMessages = new ArrayList<String>();
+                    }
+                    illegalOrphanMessages.add("You must retain Ventas " + ventasListOldVentas + " since its idVendedor field is not nullable.");
+                }
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
+            List<Ventas> attachedVentasListNew = new ArrayList<Ventas>();
+            for (Ventas ventasListNewVentasToAttach : ventasListNew) {
+                ventasListNewVentasToAttach = em.getReference(ventasListNewVentasToAttach.getClass(), ventasListNewVentasToAttach.getId());
+                attachedVentasListNew.add(ventasListNewVentasToAttach);
+            }
+            ventasListNew = attachedVentasListNew;
+            vendedores.setVentasList(ventasListNew);
+            
+            for (Ventas ventasListNewVentas : ventasListNew) {
+                if (!ventasListOld.contains(ventasListNewVentas)) {
+                    Vendedores oldIdVendedorOfVentasListNewVentas = ventasListNewVentas.getIdVendedor();
+                    ventasListNewVentas.setIdVendedor(vendedores);
+                    ventasListNewVentas = em.merge(ventasListNewVentas);
+                    if (oldIdVendedorOfVentasListNewVentas != null && !oldIdVendedorOfVentasListNewVentas.equals(vendedores)) {
+                        oldIdVendedorOfVentasListNewVentas.getVentasList().remove(ventasListNewVentas);
+                        oldIdVendedorOfVentasListNewVentas = em.merge(oldIdVendedorOfVentasListNewVentas);
+                    }
+                }
+            }
             em.getTransaction().commit();
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
@@ -72,7 +129,7 @@ public class VendedoresJpaController implements Serializable {
         }
     }
 
-    public void destroy(Integer id) throws NonexistentEntityException {
+    public void destroy(Integer id) throws IllegalOrphanException, NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -83,6 +140,17 @@ public class VendedoresJpaController implements Serializable {
                 vendedores.getId();
             } catch (EntityNotFoundException enfe) {
                 throw new NonexistentEntityException("The vendedores with id " + id + " no longer exists.", enfe);
+            }
+            List<String> illegalOrphanMessages = null;
+            List<Ventas> ventasListOrphanCheck = vendedores.getVentasList();
+            for (Ventas ventasListOrphanCheckVentas : ventasListOrphanCheck) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This Vendedores (" + vendedores + ") cannot be destroyed since the Ventas " + ventasListOrphanCheckVentas + " in its ventasList field has a non-nullable idVendedor field.");
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
             }
             em.remove(vendedores);
             em.getTransaction().commit();
